fix(debugfs): guard against negative ch offsets
The `chan` struct’s `offset` field was previously a signed long,
allowing negative values. When used in comparisons in `buf_to_channel`,
a negative offset would be cast to a large unsigned integer, causing the
bounds check to silently fail. This could allow a negative offset to
bypass validation and lead to an out-of-bounds access during operations
like `memcpy`.

This patch changes the `offset` field to `unsigned long` and updates
affected functions to use the correct types, ensuring all offset
comparisons are safe and negative values are disallowed at the type
level.

Reported-by: Bartek Piekarski <bartek.piekarski@arm.com>

+++ lib/debugfs/dev.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2021, Arm Limited. All rights reserved.
+ * Copyright (c) 2019-2025, Arm Limited. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -164,7 +164,7 @@ void channel_close(chan_t *channel)
  * dst + len.
  * It returns the actual number of bytes that were copied.
  ******************************************************************************/
-int buf_to_channel(chan_t *channel, void *dst, void *src, int nbytes, long len)
+int buf_to_channel(chan_t *channel, void *dst, void *src, size_t nbytes, long len)
 {
 	const char *addr = src;
